ems-backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-filter==23.1
django-cors-headers==4.0.0
django-redis==5.2.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
psycopg2-binary==2.9.6
python-decouple==3.8
djangorestframework-simplejwt==5.2.2
django-otp==1.1.3
pyotp==2.9.0
qrcode==7.4.2
reportlab==4.0.4
celery==5.3.0
redis==4.5.5
Pillow==10.0.0
drf-spectacular==0.26.2
python-json-logger==2.0.7
structlog==23.1.0
sentry-sdk==1.28.0
pytest==7.3.1
pytest-django==4.5.2
factory-boy==3.2.1
faker==18.11.1
requests==2.31.0
pydantic==2.0.0
gunicorn==20.1.0
whitenoise==6.4.0
boto3==1.26.131
django-storages==1.13.2
